+++ sdk/resourcemanager/machinelearningservices/armmachinelearningservices/zz_generated_workspacefeatures_client.go
@@ -54,12 +54,12 @@ func NewWorkspaceFeaturesClient(subscriptionID string, credential azcore.TokenCr
 	return client, nil
 }
 
-// List - Lists all enabled features for a workspace
+// NewListPager - Lists all enabled features for a workspace
 // If the operation fails it returns an *azcore.ResponseError type.
 // resourceGroupName - The name of the resource group. The name is case insensitive.
 // workspaceName - Name of Azure Machine Learning workspace.
 // options - WorkspaceFeaturesClientListOptions contains the optional parameters for the WorkspaceFeaturesClient.List method.
-func (client *WorkspaceFeaturesClient) List(resourceGroupName string, workspaceName string, options *WorkspaceFeaturesClientListOptions) *runtime.Pager[WorkspaceFeaturesClientListResponse] {
+func (client *WorkspaceFeaturesClient) NewListPager(resourceGroupName string, workspaceName string, options *WorkspaceFeaturesClientListOptions) *runtime.Pager[WorkspaceFeaturesClientListResponse] {
 	return runtime.NewPager(runtime.PageProcessor[WorkspaceFeaturesClientListResponse]{
 		More: func(page WorkspaceFeaturesClientListResponse) bool {
 			return page.NextLink != nil && len(*page.NextLink) > 0